requirements.txt
mysql-connector-python>=8.0.26
Pillow>=9.0.0
python-dotenv>=0.19.0
bcrypt>=3.2.0
requests>=2.26.0